processes: ask detailed report
The original idea behind the processes' plugin configuration
was that we are interested in all the processes running inside the
container.
This is indeed true, but we hit a behaviour of the collectd
plugin we don't want:

  By default, with no process matches configured, only general statistics
  is collected: the number of processes in each state and fork rate.

(source:
https://collectd.org/documentation/manpages/collectd.conf.5.shtml#plugin_processes)

So we need to list explicitely all the processes we care about.

Signed-off-by: Francesco Romani <fromani@redhat.com>